requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
drf-spectacular
django-cors-headers
drf-writable-nested